refactor: apps/front のバレル import を直接参照へ置換
