+++ apps/front/src/components/layout/AppSidebar.tsx
@@ -1,6 +1,6 @@
 import { Pen } from "lucide-react";
 import { useTranslation } from "react-i18next";
-import { NewNote } from "@/features/notes";
+import { NewNote } from "@/features/notes/components/NewNote";
 import { cn } from "@/lib/utils";
 import { Avatar, AvatarFallback } from "../ui/avatar";
 import {
