+++ apps/front/src/features/reactions/components/ReactionCount.tsx
@@ -1,7 +1,7 @@
 import { cva, type VariantProps } from "class-variance-authority";
 import { Suspense } from "react";
 import { LoadingSpinner } from "@/components/ui/loading-spinner";
-import { CustomEmoji } from "@/features/emoji";
+import { CustomEmoji } from "@/features/emoji/components/CustomEmoji";
 import { cn } from "@/lib/utils";
 
 const textSizeBarient = {
