+++ apps/front/src/features/timeline/components/MisskeyNoteContent.tsx
@@ -3,8 +3,8 @@ import type { Note } from "misskey-js/entities.js";
 import { isPureRenote } from "misskey-js/note.js";
 import { memo, useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { CustomEmojiCtx } from "@/features/emoji";
-import { MfmText } from "@/features/mfm";
+import { CustomEmojiCtx } from "@/features/emoji/contexts/CustomEmojiContext";
+import { MfmText } from "@/features/mfm/components/MfmText";
 import { NoteReplySection } from "@/features/notes/components/NoteReplySection";
 import { ReactionButton } from "@/features/reactions/components/ReactionButton";
 import { timelineSettingsAtom } from "@/features/settings/stores/timelineSettings";
