+++ apps/front/src/routes/_authed/settings.lazy.tsx
@@ -2,13 +2,11 @@ import { createLazyFileRoute } from "@tanstack/react-router";
 import { Settings as SettingsIcon } from "lucide-react";
 import { useTranslation } from "react-i18next";
 import { ScrollArea } from "@/components/ui/scroll-area";
-import {
-  ApplicationInfo,
-  ApplicationSettings,
-  DataManagement,
-  ServerInfo,
-  TimelineSettings,
-} from "@/features/settings";
+import { ApplicationInfo } from "@/features/settings/components/ApplicationInfo";
+import { ApplicationSettings } from "@/features/settings/components/ApplicationSettings";
+import { DataManagement } from "@/features/settings/components/DataManagement";
+import { ServerInfo } from "@/features/settings/components/ServerInfo";
+import { TimelineSettings } from "@/features/settings/components/TimelineSettings";
 
 export const Route = createLazyFileRoute("/_authed/settings")({
   component: SettingsPage,
