+++ apps/front/src/routes/callback/$origin/-hooks/useAuthCallback.ts
@@ -1,7 +1,7 @@
 import { useNavigate } from "@tanstack/react-router";
 import { useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { useAuth } from "@/features/auth";
+import { useAuth } from "@/features/auth/hooks/useAuth";
 import {
   PENDING_AUTH_KEY_PREFIX,
   type PeendingAuthType,
